+++ R/tt_single.R
@@ -54,6 +54,8 @@ tt_single.tidytargets <- function(
 ) {
     
     command <- substitute(command)
+
+    require_target_output(target_output)
     
     # Target script
     target_script = glue("{tt_input$initialisation$store}.R")